Minor bug fixes

+++ src/utils/cloudinary.js
@@ -9,11 +9,11 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if (!localFilePath) return null;
+        if (!localFilePath) return null;    //return null if the file is not found
 
         //upload file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
+            resource_type: "auto"   // auto automatically assigns the file type. img-> img, vid-> vid, pdf-> raw
         });
 
         console.log("File has been uploaded successfully ", response.url);
@@ -21,7 +21,11 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
         
     } catch (error) {
-        fs.unlinkSync(localFilePath);  // deletes the locally stored file when the upload to cloudinary fails.
+        if (fs.existsSync(localFilePath)) {  // checks if the file to be deleted exists.
+            fs.unlinkSync(localFilePath);   // deletes the locally stored file when the upload to cloudinary fails.
+        }  // unlinkSync is synchronous i.e. it blocks the process until the file is deleted. Bad for handling multiple users.
+        console.error("Cloudinary upload failed:", error);
+        
         return null;
     }
 };
